Refactoring strike.js
Moved many blocks to utils
Added a settings file for roles
Not sure if this even runs anymore because I had to redo the fork :)
