+++ Bot/MessageHandlers/Settings/roles.js
@@ -1,0 +1,10 @@
+// Specific roles for our server 
+// Used within strike.js 
+// Adds and removes roles semi-automatically
+
+export const ROLES = {
+    LUCKY : "902322660262957056",
+    STRIKEONE : "896026304237932565",
+    STRIKETWO : "906695092230832139",
+    BANNED : "614522888372092928",
+}
