+++ Bot/MessageHandlers/Utils/getRoleById.js
@@ -1,0 +1,2 @@
+export const GetRoleById = (msg, roleId) => 
+    msg.guild.roles.cache.find(r => r.id === roleId);
