+++ Bot/MessageHandlers/Utils/removeRoleById.js
@@ -1,0 +1,7 @@
+import { GetRoleById } from "../Utils/getRoleById";
+
+export const removeRoleById = (msg, target, roleId) => {
+    let role = GetRoleById(msg, roleId);
+    target.roles.remove(role)
+        .catch(console.error);
+}
